More refactor

+++ src/main/java/de/jpx3/intave/detect/checks/movement/physics/simulator/DefaultPoseSimulator.java
@@ -215,7 +215,7 @@ public class DefaultPoseSimulator extends PoseSimulator {
     double resetMotion = movementData.resetMotion();
     double jumpUpwardsMotion = movementData.jumpMotion();
 
-    double interpolations = 0;
+    int interpolations = 0;
     double interpolateX = context.motionX;
     double interpolateY = context.motionY;
     double interpolateZ = context.motionZ;
